Added support to execute bash commands with quotes and other operators like &&, etc. to CrosRunCommand().
It now creates a unique file and dumps all the commands to the file,
copies the file over to the remote, and finally executes that file using
bash on the remote.

Also added command_timeout to CrosRunCommand().

This is a pre-requisite to some of the scripts I'll be adding later for
automatic profiling.

PRESUBMIT=passed
R=shenhan,bjanakiraman
DELTA=24  (21 added, 0 deleted, 3 changed)
OCL=57623-p2
RCL=57652-p2
RDATE=2012/01/04 17:19:20


P4 change: 42662792