Changed the chromiumos gcc.git repository in the following way: 1. The "master" branch now holds the sources for the current released gcc. 2. Upstream svn google/main is now tracked by a branch called gcc.gnu.org/branches/google/main. 3. All gcc branches now hold the sources in ./ instead of 2 directories down. This simplies things in the ebuild and it doesn't have to worry about differences in directories like: gcc/4.6.0 vs. gcc/4.6.2.
This change is needed because we will soon be shifting to cros_workon
toolchain packages. That means a single branch ("master" for us), will
hold the sources for the currently released compiler.
When that branch is updated, an automatic uprev of the gcc package will
occur.

PRESUBMIT=passed
R=bjanakiraman,shenhan,yunlian
DELTA=15  (13 added, 0 deleted, 2 changed)
OCL=58311-p2
RCL=58331-p2
RDATE=2012/02/09 15:48:37


P4 change: 42667272