Removed logging of cached output and error to the console.
I noticed minutes of time are wasted on printing the cached output to
the console especially if the number of runs is large. In one instance,
my machine was under heavy load for 5+ minutes when it was producing
output for ~140 completed autotests.

This CL causes crosperf to no longer print on the console the cached run
output.

It also causes the stderr to go to the *.err file properly.

Tested: Ran crosperf with some cached results and found it much faster
than before.

PRESUBMIT=passed
R=yunlian,bjanakiraman
APPROVED=yunlian
DELTA=2  (1 added, 0 deleted, 1 changed)
OCL=63443-p2
RCL=63540-p2
RDATE=2012/10/22 11:47:35


P4 change: 42796921