Disable ccache by default in the sysroot wrapper
We have found that ccache ignores GCC envvars and may not be
checking for diffs in all the GCC binaries. It is safer for
the toolchain team to disable this.
Moreover, in my experiments, ccache seems to make the builds
slower when using a large parallel factor.
There is a new option to enable ccache.

TEST=Build with and without the new option to enable ccache and
verified that sysroot wrapper is correct.

PRESUBMIT=passed
R=cmtice,asharif,shenhan
APPROVED=shenhan
DELTA=16 (13 added, 0 deleted, 3 changed)
OCL=43620049


P4 change: 43669814