Missed another instance of return without "finally" and no "except problem in this file.
Fixed location of return statement. Should not be within "finally" block.

When return statement is within "finally" statement any exception
that happens within the try block will not be reported. Exceptions are
supposed to be rethrown at the end of "finally" if there was no "except"
but, as the code was written, the end of "finally" is never reached.

TEST=Verified exceptions are reported correctly. Also, with asharif
verified with small test case that our understanding of "finally" was
correct.

PRESUBMIT=passed
R=asharif
APPROVED=asharif
DELTA=1 (0 added, 0 deleted, 1 changed)
OCL=43793218


P4 change: 43802609