Merge the test stage and the build stage.
Use the pipeline_worker to represent both stage. These two stages differ on
getting the key of the task in the stage and setting the value of this stage.
The key uniquely distinguish a task from another task. For example, the key
for the build phase is the optimization flags, while the key for the test stage
is the checksum of the built image.

BUG=None
TEST=None

Change-Id: Id057c17026806acd406fe78cc70ed996273ca9aa
Reviewed-on: https://gerrit-int.chromium.org/40376
Reviewed-by: Simon Que <sque@google.com>
Reviewed-by: Luis Lozano <llozano@chromium.org>
Commit-Queue: Yuheng Long <yuhenglong@google.com>
Tested-by: Yuheng Long <yuhenglong@google.com>