Check for file before prepending directory path.
Sometimes _GetDataMeasurementsFiles returns the file with the full
directory path, and sometimes it doesn't.  We were assuming it always
did not.  This patch checks to see if we can already find the file
before attempting to prepend the directory path.

BUG=None
TEST=Tested this in the nightly infrastructure tests area, where we
were seeing the problem.  This fixed that problem.  (Before this fix
we were not seeing the successful test results; after it we are.)

Change-Id: Ic0a2df4c9aff7b820f652f7ecad56976ed97a4b0
Reviewed-on: https://chrome-internal-review.googlesource.com/149256
Reviewed-by: Luis Lozano <llozano@chromium.org>
Commit-Queue: Caroline Tice <cmtice@google.com>
Tested-by: Caroline Tice <cmtice@google.com>