Use default results for Telemetry tests; add option to use all results.
When running telemetry_Crosperf, use the json file of report defaults to
determine which results to actually include in the report.  This CL
also adds a flag to Crospref, --show_all_results=True, that allows the
user to override the defaults and generate a report showing all the values.

BUG=None
TEST=Tested this both with and without the --show_all_results flag. It
works as expected for me.

Change-Id: I91364572ebdf3d54bb8b408fbd5a2580923f6829
Reviewed-on: https://chrome-internal-review.googlesource.com/149457
Reviewed-by: Yunlian Jiang <yunlian@google.com>
Commit-Queue: Caroline Tice <cmtice@google.com>
Tested-by: Caroline Tice <cmtice@google.com>