Update image_chromeos.py to use 'cros flash'
Check the chroot to see if it contains the 'cros flash' script; if so
use 'cros flash' instead of 'cros_image_to_target.py' for imaging the
DUT.  This is the first of a series of changes to improve the imaging
options in Crosperf.  This script does not change any of the options or
syntax for using image_chromeos.py.

BUG=None
TEST=Tested in chroot with & without cros_flash.py.  Worked correctly in
both cases.

Change-Id: If8704c81aa800cab85379baa514249105fe2b83f
Reviewed-on: https://chrome-internal-review.googlesource.com/153197
Reviewed-by: Han Shen <shenhan@google.com>
Commit-Queue: Caroline Tice <cmtice@google.com>
Tested-by: Caroline Tice <cmtice@google.com>