Minor Crosperf fixes and cleanups.
Fix crosperf so that it doesn't try to look for a default result
(from the json file) if the telemetry test failed altogether.  Also
fix it so that if it can't find the default result (on a successful run),
it returns all the results, rather than exiting with a backtrace.  Also
add units to results.

BUG=None
TEST=Tested these fixes on my system.  They appear to work as desired now.

Change-Id: I86099a1174c4d91d7d92a5aeea0792c68fa8dbbb
Reviewed-on: https://chrome-internal-review.googlesource.com/154114
Reviewed-by: Yunlian Jiang <yunlian@google.com>
Commit-Queue: Caroline Tice <cmtice@google.com>
Tested-by: Caroline Tice <cmtice@google.com>