Update to use "working" version of perf for now.
This is a temporary fix, until we can figure out why the 'normal' version
of 'perf' is not properly symbolizing ChromeOS perf data.  This CL
commits a version of the 'perf' binary (based on perf 3.13) that does
properly symbolize ChromeOS perf data files.  It also updates crosperf to
copy this working perf binary into the chroot (in /tmp) and to use it
when generating the perf report.  Work to track down and fix the real
problem is ongoing, but in the meantime this allows Crosperf users to
get meaningful perf reports for ChromeOS.

BUG=None
TEST=Tested Crosperf using this patch and generating perf reports.

Change-Id: Ia8dd157b88d37a897eb9becf8fa444de41e88873
Reviewed-on: https://chrome-internal-review.googlesource.com/157756
Reviewed-by: Caroline Tice <cmtice@google.com>
Commit-Queue: Caroline Tice <cmtice@google.com>
Tested-by: Caroline Tice <cmtice@google.com>