Remove 'Completed' column; Remove 'retval' row from summary table.
The "Completed" column causes confusion because tests that fail are
still considered to have completed.  This CL removes the 'Completed'
column, and instead posts the number of iterations that passed or failed,
for each test image, in the table header, beside the image name.
It also removes the 'retval' row from the summary table, since the retval
will either be '0' if a test iteration passes or '1' if it fails, and seeing
the average of all the retvals is useless, unless they all pass or fail anyway.
The retval row is still there in the full table, which shows all the results
for each iteration.

BUG=357340
TEST=I ran crosperf with this change in all configurations I could think of and
it seems to work.

Change-Id: I0e64b9ed016b0a4604814bac49442848cd27f5d3
Reviewed-on: https://chrome-internal-review.googlesource.com/159113
Reviewed-by: Yunlian Jiang <yunlian@google.com>
Commit-Queue: Caroline Tice <cmtice@google.com>
Tested-by: Caroline Tice <cmtice@google.com>