Always write verbose logs to files.
The option that reduces the amount of screen output (--logging_level
= 'quiet' or 'average') was also reducing what got written to the log
files.  This fixes that.  With this CL, even though the amount of screen
output is reduced, what gets written to the log files is always the same
(verbose).  This CL also makes "average" the default logging level.

BUG=None
TEST=I tested all three logging levels, saving the results in different
log directories and compared the log files afterwards, to make sure
the log files contained the same amount of logging, regardless of the
screen logging level.

Change-Id: I6982499d67c080afa835e01ab4fa2282dd992052
Reviewed-on: https://chrome-internal-review.googlesource.com/159654
Reviewed-by: Caroline Tice <cmtice@google.com>
Tested-by: Caroline Tice <cmtice@google.com>
Commit-Queue: Caroline Tice <cmtice@google.com>