Better handling of "not identical machines" failure.
Nightly tests, especially on the x86-generic boxes, fail frequently
because the DUTs have different board images on them, so Crosperf
decides the machines are not identical and refuses to run the tests.
With this CL, if the machine_manager finds that the machines fail the
"identical" test, it will try to push the same image onto all the machines,
and then check them again to see if they are the same.  It only tries this
once; if they fail the check the second time around, it is still a fatal
failure.  This should eliminate many of the unnecessary failures in our
nightly tests.

This CL also fixes a small bug  in the auto-delete script (this fix has
been running for a while on mobiletc-prebuild, but never got committed).

BUG=None
TEST=Ran several iterations where I forced the first "identical" check to
fail.  The changes worked.

Change-Id: Ied2a55e5d3e2789e58a503aef03269888954b579
Reviewed-on: https://chrome-internal-review.googlesource.com/163334
Reviewed-by: Luis Lozano <llozano@chromium.org>
Commit-Queue: Caroline Tice <cmtice@google.com>
Tested-by: Caroline Tice <cmtice@google.com>