Add --skip_chroot_upgrade flag to build_packages.
This CL adds the '--skip_chroot_upgrade' flag to all calls to
build_packages that clobber CFLAGS, LDFLAGS, etc.  This is to
prevent the call to build_packages from deciding to try to rebuild
the compiler itself (which is not the right place to do that).

BUG=None
TEST=Tested adding flag by hand to call to build_packages that used to
fail; it passed with the flag.

Change-Id: Ia0152f76f8a25cf0488a4a72c4add7e4b3182feb
Reviewed-on: https://chrome-internal-review.googlesource.com/164050
Reviewed-by: Han Shen <shenhan@google.com>
Commit-Queue: Caroline Tice <cmtice@google.com>
Tested-by: Caroline Tice <cmtice@google.com>