Add --cache_only global setting to crosperf.
The CL adds a new setting, --cache_only, which causes crosperf to
ONLY report test results that it finds in the cache (it will NOT
rerun any tests, even if no cache results are found).  This is partly
for implementing a weekly summary report, where we want to compare the
nightly results from the past week, but not actually rerun anything.

BUG=None
TEST=I tested this option while generating weekly test reports, and it
worked as it is supposed to.

Change-Id: Icb1f5343f76ce999bebd0810bb86ab64a7577313
Reviewed-on: https://chrome-internal-review.googlesource.com/164890
Reviewed-by: Yunlian Jiang <yunlian@google.com>
Commit-Queue: Caroline Tice <cmtice@google.com>
Tested-by: Caroline Tice <cmtice@google.com>