Update test_toolchains.py to save tar files of the images used.
This is for generating a weekly report that compares the nightly
test results.  All of the nightly test results are cached, but in
order to find the correct cache results, you need the image from
which the results were created.  Due to disk space, some of the
chroots will be deleted before the weekly report gets generated.
Therefore, this modification tars up the test & vanilla image
directories and stores them separately, where they can be accessed
for generating the weekly report.

BUG=None
TEST=Used this successfully last night for nightly test runs.

Change-Id: I01b457a402c3649d7904a478b37bad60b3395300
Reviewed-on: https://chrome-internal-review.googlesource.com/164879
Reviewed-by: Yunlian Jiang <yunlian@google.com>
Commit-Queue: Caroline Tice <cmtice@google.com>
Tested-by: Caroline Tice <cmtice@google.com>