Update nightly tests (and weekly report stuff) to use official builds.
This CL changes the nightly tests to download and use the official
image for our 'vanilla' image, rather than building the vanilla image
ourselves.  It also updates the code that creates & copies the tar
files of the images, for the weekly reports, to look for and copy the
official images.

BUG=None
TEST=Tested the changes by hand.  Also let them run for a couple of
nights for our nightly tests.

Change-Id: I8c3215aab7550dc9f148ed79095295f702141017
Reviewed-on: https://chrome-internal-review.googlesource.com/166477
Reviewed-by: Luis Lozano <llozano@chromium.org>
Commit-Queue: Caroline Tice <cmtice@google.com>
Tested-by: Caroline Tice <cmtice@google.com>