Fix issue where Telemetry tests appear to pass, but really failed
Occasionally Telemetry tests will return a "pass" even though the test
did not run (e.g. if the test was disabled).  Crosperf was looking for
results for every passing test and dying when it failed to find a
result for these false passes.  This CL looks for the false pass and
changes it fail, as well as writing out a warning message.

A bug was filed for this in Buganizer:
https://b.corp.google.com/issue?id=16190266, but since this was not
filed as a Chrome/ChromeOS bug, I will leave the bug mark below as
None (to avoid confusing the bug droid).

BUG=None
TEST=Tested this against a Telemetry run with both false and real
passes, and it worked properly.

Change-Id: Ie010c0cd5b0775d0e20cfd6967975eb4e768c55c
Reviewed-on: https://chrome-internal-review.googlesource.com/169115
Reviewed-by: Yunlian Jiang <yunlian@google.com>
Commit-Queue: Caroline Tice <cmtice@google.com>
Tested-by: Caroline Tice <cmtice@google.com>