Add AFDO to nightly builds.
This adds an option, '--use_afdo' to build_chromeos.py, along with
a list of boards for which it is valid. The flag will cause
build_chromeos.py to add 'USE=afdo_use' to calls to build_packages,
if the board is appropriate. This CL also updates test_toolchains.py
to pass --use_afdo to build_chromeos, if appropriate.

BUG=None
TEST=Tried it in my own directory and it worked.

Change-Id: I0a2056f86783dbd30a2fd650729c888fd1df1f67
Reviewed-on: https://chrome-internal-review.googlesource.com/169174
Reviewed-by: Luis Lozano <llozano@chromium.org>
Commit-Queue: Caroline Tice <cmtice@google.com>
Tested-by: Caroline Tice <cmtice@google.com>