New script to tell if machines are 'identical' for crosperf.
This script calls the functions in machine_manager.py and
verifies that two or more machines are 'identical' for the purposes
of running crosperf.  I have several times needed to do this type
of thing, so I decided to extract the functionality into its own
script.

BUG=None
TEST=Tested in my directory.

Change-Id: I2a410eefe7e1b46f6c5cad477640ea7974a614c7
Reviewed-on: https://chrome-internal-review.googlesource.com/188197
Reviewed-by: Luis Lozano <llozano@chromium.org>
Reviewed-by: Caroline Tice <cmtice@google.com>
Commit-Queue: Caroline Tice <cmtice@google.com>
Tested-by: Caroline Tice <cmtice@google.com>