Search Google Storage for trybot image artifacts.
Sometimes trybot jobs fail in the testing phase, but the build
finished and a build artifact exists.  To deal with this, instead
of checking the trybot job status, we just search for the build
artifact in google storage.  This also removes the
'--chrome_rev=tot' flag from the buildbot command.

Change-Id: I38c07270d774258953d9b1f1e694aba9730ac92b
Reviewed-on: https://chrome-internal-review.googlesource.com/197366
Reviewed-by: Luis Lozano <llozano@chromium.org>
Tested-by: Caroline Tice <cmtice@google.com>
Commit-Queue: Caroline Tice <cmtice@google.com>