Use original autotest location for telemetry runs.
For telemetry benchmarks, there is not need for a board specific
version of autotest. There are no compiled tests. So it is easier to
use the "source" location of autotest. This avoid having to build
autotest for the specific board which can be time consuming.
This also fixes the problem where after doing a build, sometimes
the telemetry_Crosperf test cannot be found since there is no need
for the copy.

This does nothing regarding autotest compiled tests. These still need
to be under /build/<board>. If using downloaded image, we could also
download the auotest bits but we will probably need a new "mode" or
flag for crosperf to indicate we want to use the autotest bits that
come from the image and not the one in your chroot. So, more work is
required.

BUG=None
TEST=Tested by hand with multiple telemetry benchmarks.

Change-Id: Id669782a9612fad39326b4d820cbc8e78a8a3d0c
Reviewed-on: https://chrome-internal-review.googlesource.com/209980
Reviewed-by: Caroline Tice <cmtice@google.com>
Commit-Queue: Luis Lozano <llozano@chromium.org>
Tested-by: Luis Lozano <llozano@chromium.org>