Fix current problem with machine locks.
Currently if crosperf can't access the default locks directory, it
assumes that it is safe to proceed without locks and it uses machines
without checking locks.  This script does two things:  It causes
crosperf to fail if it can't access a locks directory; and it allows
the user to specify an alternate locks directory.  As a by-product it
also adds a boolean prompting function to utils/misc.py

BUG=None
TEST=Tested with default directory accessible & inaccessible; tested using
local locks directory.

Change-Id: I6f24f772986813bf089f5f4e40335a15e8ee4837
Reviewed-on: https://chrome-internal-review.googlesource.com/214129
Reviewed-by: Luis Lozano <llozano@chromium.org>
Commit-Queue: Caroline Tice <cmtice@google.com>
Tested-by: Caroline Tice <cmtice@google.com>