Update nightly tests to use mail-sheriff program for reports.
This updates the nightly buildbot and local tester programs to
tell crosperf to put its results in a special directory, and to
not send email.  After crosperf finishes, it looks for the email
information in the special directory and sends out then email using
the mail-sheriff program.

BUG=None
TEST=Currently running buildbot test.  Will not commit the change
until the test passes.  Since the test takes several hours to run
I wanted to put this out for review now.

Change-Id: Idd3d137e4a6e144b677690fc6775580dfac56af1
Reviewed-on: https://chrome-internal-review.googlesource.com/217286
Reviewed-by: Luis Lozano <llozano@chromium.org>
Commit-Queue: Caroline Tice <cmtice@google.com>
Tested-by: Caroline Tice <cmtice@google.com>