Update nightly tests to use old file locking mechanism for now.
There are some issue with the role account using the AFE server
locking mechanism, so we will temporarily revert the nightly tests
to use the old file locking mechanism.  This CL depends on the CL
that re-enables the file locking mechanism (with a new flag).

BUG=None
TEST=None
CQ-DEPEND=CL:219096

Change-Id: If525e4fcf292d3c57aa9c69a1b796914bd6d434c
Reviewed-on: https://chrome-internal-review.googlesource.com/219088
Reviewed-by: Yunlian Jiang <yunlian@google.com>
Commit-Queue: Caroline Tice <cmtice@google.com>
Tested-by: Caroline Tice <cmtice@google.com>