Remove temp file that breaks telemetry.
Occasionally after running Telemetry, it will leave behind a
file in /tmp of your chroot (/tmp/telemetry@<machine_name>), which
will cause any later attempt to run telemetry on that machine to
fail, until the file is removed.

This CL checks for such a file, and if found deletes it, before
attempting to run the telemetry tests.

BUG=chromium:513875
TEST=Tested in my chroot with & without such a file.

Change-Id: I3e0a3000a1614f46f9312abddf77fd48b25334da
Reviewed-on: https://chrome-internal-review.googlesource.com/223812
Reviewed-by: Luis Lozano <llozano@chromium.org>
Commit-Queue: Caroline Tice <cmtice@google.com>
Tested-by: Caroline Tice <cmtice@google.com>