Make duplicate locking/unlocking warning instead of error.
Change the afe_lock_manager, so that an attempt to lock an already
locked machine or unlock an already unlocked machine is a warning
rather than an exception; also return a list of machines which we
were successfully able to lock/unlock.

Also update Crosperf: Previously if multiple machines were requested
and we were not able to lock all of them, the Crosperf job failed.
Now if we are able to successfully lock any of the requested machines,
Crosperf will try to run the job.

BUG=None
TEST=Tested in my workspace with various locked/unlocked
configurations for crosperf job with two machines.

Change-Id: I65a0a9cce4b8be8d0c043913dc2dd7f9f07cda0f
Reviewed-on: https://chrome-internal-review.googlesource.com/224180
Reviewed-by: Han Shen <shenhan@google.com>
Commit-Queue: Caroline Tice <cmtice@google.com>
Tested-by: Caroline Tice <cmtice@google.com>