Temporarily use good external Chrome source tree.
This is to work around a problem with the current version of Chrome
inside ChromeOS (telemetry is broken in that version of Chrome).
This CL makes the nightly tests use an external Chrome source tree
that has been verified to work properly.  This will be un-done, once
ChromeOS gets a good version of Chrome.

BUG=None
TEST=Tested crosperf on chrotomation2 using this.

Change-Id: I41caef9bb1b1973c83745e3962a5c36c489d7514
Reviewed-on: https://chrome-internal-review.googlesource.com/226695
Reviewed-by: Luis Lozano <llozano@chromium.org>
Commit-Queue: Luis Lozano <llozano@chromium.org>
Tested-by: Luis Lozano <llozano@chromium.org>