Add json report archiving.
This CL adds a new output, in json format, for archiving test
results.  It also adds a new flag, --json_report, to crosperf to
tell it to generate the json output file in the usual results
directory.  The json output is always in addition to the other
reports that Crosperf generates.  This CL also make some minor
changes to the tool that maintains the telemetry default results
file, so it can be used by the json report generator.

BUG=None
TEST=Tested Crosperf with & without new flag; it did what it should.

Change-Id: Id0e476716ceee208d33e6f9427ea95235fbf11d4
Reviewed-on: https://chrome-internal-review.googlesource.com/228325
Reviewed-by: Han Shen <shenhan@google.com>
Commit-Queue: Caroline Tice <cmtice@google.com>
Tested-by: Caroline Tice <cmtice@google.com>