Various small cache cleanups.
Add the machine used for running the test to benchmark run; use the
actual data from that machine for generating the cache entry for the
run.  Also make sure, when calculating the common checksums for looking
machines up in the cache, that all the machines being used actually have
the same checksum.  Also add a cache_keys.txt file to the cache directory,
to allow people to see what was used for generating the cache
entry/checksums.

BUG=None
TEST=Tested in my directory.

Change-Id: Ia4cf8316e920a24becf5d12fd4c8f9a3e30a180f
Reviewed-on: https://chrome-internal-review.googlesource.com/228854
Commit-Ready: Caroline Tice <cmtice@google.com>
Tested-by: Caroline Tice <cmtice@google.com>
Reviewed-by: Yunlian Jiang <yunlian@google.com>