Filter out unreachable machines from experiment.remote as well as label.remote.
We had a bug for a experiment file with label L1 and L2, remote A & B,
and B is unreachable (dead machine). This ends up with
MachineManager._all_machines being [A], but L1.remote and L2.remote
being both [A, B], because B is locked even it is not reachable. This
caused an internal error because only 1 dut worker is create for 'A', no
dut worker created for 'B'.

Fixed by filter unreachable machines from both experiment.remote as well
as label.remote.

Tested by running a experiment file manually.

Change-Id: Ia3f563fe7e6d1521ec8d506b79cdf185af76df06
Reviewed-on: https://chrome-internal-review.googlesource.com/231038
Commit-Queue: Han Shen <shenhan@google.com>
Tested-by: Han Shen <shenhan@google.com>
Reviewed-by: Han Shen <shenhan@google.com>