Temporarily remove graphics_WebGLAquarium from buildbot testing.
For the images that we build in trybots, this test cannot be
executed because we are not downloading the necessary artifacts
for autotest. See chromium:551710.
This change temporarily removes that test from the nightly testing that
we do based on trybot images. These changes should be removed
once this bug is fixed.

BUG=chromium:551710
TEST=None

Change-Id: I15ec9386d52809492345d9cd46a6f0ad39330dbc
Reviewed-on: https://chrome-internal-review.googlesource.com/237838
Commit-Ready: Luis Lozano <llozano@chromium.org>
Tested-by: Luis Lozano <llozano@chromium.org>
Reviewed-by: Caroline Tice <cmtice@google.com>