Add a new shell script to figure out which compiler built image.
This adds a new script that checks to see which compiler (GCC or
clang) was used to build all the images in a ChromeOS image.  It
outputs a count of how many packages it found that were built with
each compiler.

BUG=None
TEST=Tested in my chroot with both a locally-built GCC image and a
trybot-built clang image.

Change-Id: I26b5faea8616b5427c5bbe27a58ea3c9d4e36043
Reviewed-on: https://chrome-internal-review.googlesource.com/238685
Commit-Ready: Caroline Tice <cmtice@google.com>
Tested-by: Caroline Tice <cmtice@google.com>
Reviewed-by: Han Shen <shenhan@google.com>