Display '.cros' at end of lab machine names; add it when missing.
Before this, this script was confusing because it would display
lab machine names without the '.cros' at the end, but it would require
the '.cros' at the end of lab machine names (from users) to recognize
them.  This CL fixes that:  Now when displaying machine status, it
makes sure the .cros is at the end of the lab machine names; and it
makes sure the various functions work for lab machines whether or not
the user specified the '.cros' at the end of the machine name (it adds
it automatically, if it's missing).

BUG=chromium:525198
TEST=Tested all the commands on lab machines with and without '.cros'.

Change-Id: I8fd36d116c3c542fad936be73b946c88a9228551
Reviewed-on: https://chrome-internal-review.googlesource.com/240880
Commit-Ready: Caroline Tice <cmtice@google.com>
Tested-by: Caroline Tice <cmtice@google.com>
Reviewed-by: Yunlian Jiang <yunlian@google.com>