Fix arguments to RunCommand from RunCommands.
Under the recent RunCommand changes, all the arguments to RunCommand
except the command itself must be keyword args.  The initial change
missed fixing this in the RunCommands function, which resulted in
an assertion error in the overnight test.  This CL fixes that issue.

BUG=None
TEST=Tested running setup_chromeos with the role account (where
previous failur occurred).

Change-Id: I73834309afc5cdbddd8953de5287bf298fb80b7e
Reviewed-on: https://chrome-internal-review.googlesource.com/241980
Commit-Ready: Caroline Tice <cmtice@google.com>
Tested-by: Caroline Tice <cmtice@google.com>
Reviewed-by: Luis Lozano <llozano@chromium.org>