toolchain-utils: strip argv[0] before calling Main().
https://chrome-internal-review.googlesource.com/#/c/244221 fixed some
cros lint errors, and in the process, modified test_toolchains.py to
use argparse instead of optparse.

The flags parsing semantics of these modules seem to differ when they
see an argument which is not a flag. As a result, "nightly_wrapper.sh"
nightly cron jobs are failing with error: "unrecognized arguments:
  /usr/local/google/crostc/toolchain-utils/test_toolchains.py"

This CL strips the argv[0] before calling Main(), since the name of
the script is not really an argument and doesn't need to be parsed.

BUG=None
TEST='nightly_wrapper.sh lumpy ...' proceeds with no error during argument parsing.

Change-Id: Idf13ab457170bb051d4d0693a887795d0c16918d
Reviewed-on: https://chrome-internal-review.googlesource.com/245868
Commit-Ready: Luis Lozano <llozano@chromium.org>
Tested-by: Luis Lozano <llozano@chromium.org>
Reviewed-by: Luis Lozano <llozano@chromium.org>