[crosperf] Fix default result value for media.tough_video_cases.
The telemetry test media.tough_video_cases does not always
generate the result value that the defaults file was looking
for.  This causes crosperf to fail entirely on this test sometimes.
Also, the old default value is no longer the "value of interest"
for that test.  This CL removes the no-longer-valid default values.

BUG=chromium:585136
TEST=None

Change-Id: I484408f54dcac29efbec9879147d9490a3b17eba
Reviewed-on: https://chrome-internal-review.googlesource.com/248045
Commit-Ready: Caroline Tice <cmtice@google.com>
Tested-by: Caroline Tice <cmtice@google.com>
Reviewed-by: Yunlian Jiang <yunlian@google.com>